Added authorization to user requests.

+++ backend/routes/posts.js
@@ -7,7 +7,8 @@ const router = express.Router();
 router.post("", checkAuth, (req, res, next) => {
     const post = new Post({
         title: req.body.title,
-        content: req.body.content
+        content: req.body.content,
+        creator: req.userData.userId
     });
     post.save().then(createdPost => {
         res.status(201).json({
@@ -23,8 +24,12 @@ router.put("/:id", checkAuth, (req, res, next) => {
         title: req.body.title,
         content: req.body.content
     });
-    Post.updateOne({ _id: req.params.id }, post).then(result => {
-        res.status(200).json({ message: "Update successful!" });
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+        if (result.n > 0) {
+            res.status(200).json({ message: "Update successful!" });
+        } else {
+            res.status(401).json({ message: "Unauthorized PUT request." });
+        }
     });
 });
 
@@ -48,9 +53,12 @@ router.get("/:id", checkAuth, (req, res, next) => {
 })
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-    Post.deleteOne({ _id: req.params.id }).then(result => {
-        console.log(result);
-        res.status(200).json({ message: "Post deleted!" });
+    Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+        if (result.nModified > 0) {
+            res.status(200).json({ message: "Delete request successful!" });
+        } else {
+            res.status(401).json({ message: "Unauthorized DELETE request." });
+        }
     });
 });
